requirements.txt
Django==5.0.2
djangorestframework==3.14.0
markdown==3.5.2
django-filter==23.5
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
Pillow==10.2.0
django-agora==0.4
django-channels==0.7.0
razorpay==1.4.1
Pyrebase4==4.7.1
django-unfold==0.20.3